Added incomplete unit tests for the Math library, as well as some bug-fixes discovered while writing it.
These tests cover the entirety of the Matrix2 and Vector2 class, although they aren't templated yet. Unit tests are done with the Google Test framework and require Visual Studio 2015 to build. The included project and solution should work on every system and include the Google Test library and header. The tests can be used as a basis for testing other parts of the project such as containers.